LPS-51883 Eliminate PACLIntegrationJUnitTestRunner
