+++ portal-test-internal/src/com/liferay/portal/test/PACLTestRule.java
@@ -15,6 +15,9 @@
 package com.liferay.portal.test;
 
 import com.liferay.portal.deploy.hot.HookHotDeployListener;
+import com.liferay.portal.deploy.hot.IndexerPostProcessorRegistry;
+import com.liferay.portal.deploy.hot.SchedulerEntryRegistry;
+import com.liferay.portal.deploy.hot.ServiceWrapperRegistry;
 import com.liferay.portal.kernel.deploy.hot.DependencyManagementThreadLocal;
 import com.liferay.portal.kernel.deploy.hot.HotDeployEvent;
 import com.liferay.portal.kernel.deploy.hot.HotDeployUtil;
@@ -22,18 +25,38 @@ import com.liferay.portal.kernel.portlet.PortletClassLoaderUtil;
 import com.liferay.portal.kernel.servlet.ServletContextPool;
 import com.liferay.portal.kernel.servlet.filters.invoker.InvokerFilterHelper;
 import com.liferay.portal.kernel.test.BaseTestRule;
-import com.liferay.portal.kernel.test.ReflectionTestUtil;
 import com.liferay.portal.kernel.util.ClassLoaderPool;
+import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.PortalLifecycleUtil;
+import com.liferay.portal.kernel.util.PropsKeys;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.service.ServiceTestUtil;
 import com.liferay.portal.spring.context.PortletContextLoaderListener;
 import com.liferay.portal.test.mock.AutoDeployMockServletContext;
-import com.liferay.portal.test.runners.PACLIntegrationJUnitTestRunner;
+import com.liferay.portal.util.InitUtil;
 import com.liferay.portal.util.PortalUtil;
+import com.liferay.portal.util.PropsUtil;
+
+import java.lang.reflect.Method;
+
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
+
+import java.security.CodeSource;
+import java.security.ProtectionDomain;
+
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+import javax.naming.Context;
 
 import javax.servlet.ServletContext;
 import javax.servlet.ServletContextEvent;
 
 import org.junit.runner.Description;
+import org.junit.runners.model.Statement;
 
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.FileSystemResourceLoader;
@@ -47,7 +70,8 @@ import org.springframework.mock.web.MockServletContext;
  */
 public class PACLTestRule extends BaseTestRule<HotDeployEvent, Object> {
 
-	public static final PACLTestRule INSTANCE = new PACLTestRule();
+	public static final String RESOURCE_PATH =
+		"com/liferay/portal/security/pacl/test/dependencies";
 
 	@Override
 	protected void afterClass(
@@ -75,10 +99,11 @@ public class PACLTestRule extends BaseTestRule<HotDeployEvent, Object> {
 	}
 
 	@Override
-	protected HotDeployEvent beforeClass(Description description) {
-		ReflectionTestUtil.setFieldValue(
-			description, "fTestClass",
-			PACLIntegrationJUnitTestRunner.getCurrentTestClass());
+	protected HotDeployEvent beforeClass(Description description)
+		throws ReflectiveOperationException {
+
+		_testClass = _loadTestClass(description.getTestClass());
+		_instance = _testClass.newInstance();
 
 		ServletContext servletContext = ServletContextPool.get(
 			PortalUtil.getServletContextName());
@@ -101,9 +126,7 @@ public class PACLTestRule extends BaseTestRule<HotDeployEvent, Object> {
 
 		PortalLifecycleUtil.flushInits();
 
-		Class<?> testClass = description.getTestClass();
-
-		ClassLoader classLoader = testClass.getClassLoader();
+		ClassLoader classLoader = _testClass.getClassLoader();
 
 		MockServletContext mockServletContext = new MockServletContext(
 			new PACLResourceLoader(classLoader));
@@ -153,7 +176,189 @@ public class PACLTestRule extends BaseTestRule<HotDeployEvent, Object> {
 		}
 	}
 
-	private PACLTestRule() {
+	@Override
+	protected void invokeStatement(Statement statement, Description description)
+		throws Throwable {
+
+		String methodName = description.getMethodName();
+
+		if (methodName == null) {
+			statement.evaluate();
+
+			return;
+		}
+
+		Method method = _testClass.getMethod(description.getMethodName());
+
+		method.invoke(_instance);
+	}
+
+	private static Class<?> _loadTestClass(Class<?> clazz)
+		throws ClassNotFoundException {
+
+		ProtectionDomain protectionDomain = clazz.getProtectionDomain();
+
+		CodeSource codeSource = protectionDomain.getCodeSource();
+
+		ClassLoader classLoader = new PACLClassLoader(
+			new URL[] {codeSource.getLocation()}, clazz.getClassLoader());
+
+		return Class.forName(clazz.getName(), true, classLoader);
+	}
+
+	private static final String _PACKAGE_PATH =
+		"com.liferay.portal.security.pacl.test.";
+
+	static {
+		URL resource = PACLTestRule.class.getResource("pacl-test.properties");
+
+		if (resource != null) {
+			System.setProperty("external-properties", resource.getPath());
+		}
+
+		System.setProperty(
+			Context.INITIAL_CONTEXT_FACTORY,
+			"org.apache.naming.java.javaURLContextFactory");
+
+		System.setProperty("catalina.base", ".");
+
+		List<String> configLocations = ListUtil.fromArray(
+			PropsUtil.getArray(PropsKeys.SPRING_CONFIGS));
+
+		InitUtil.initWithSpring(configLocations, true);
+
+		ServiceTestUtil.initServices();
+		ServiceTestUtil.initPermissions();
+
+		new IndexerPostProcessorRegistry();
+		new SchedulerEntryRegistry();
+		new ServiceWrapperRegistry();
+	}
+
+	private Object _instance;
+	private Class<?> _testClass;
+
+	private static class PACLClassLoader extends URLClassLoader {
+
+		public PACLClassLoader(URL[] urls, ClassLoader parentClassLoader) {
+			super(urls, parentClassLoader);
+		}
+
+		@Override
+		public URL findResource(String name) {
+			if (_urls.containsKey(name)) {
+				return _urls.get(name);
+			}
+
+			URL resource = null;
+
+			if (!name.contains(RESOURCE_PATH)) {
+				String newName = name;
+
+				if (!newName.startsWith(StringPool.SLASH)) {
+					newName = StringPool.SLASH.concat(newName);
+				}
+
+				newName = RESOURCE_PATH.concat(newName);
+
+				resource = super.findResource(newName);
+			}
+
+			if ((resource == null) && !name.contains(RESOURCE_PATH)) {
+				String newName = name;
+
+				if (!newName.startsWith(StringPool.SLASH)) {
+					newName = StringPool.SLASH.concat(newName);
+				}
+
+				newName = RESOURCE_PATH.concat("/WEB-INF/classes").concat(
+					newName);
+
+				resource = super.findResource(newName);
+			}
+
+			if (resource == null) {
+				resource = super.findResource(name);
+			}
+
+			if (resource != null) {
+				_urls.put(name, resource);
+			}
+
+			return resource;
+		}
+
+		@Override
+		public URL getResource(String name) {
+			if (name.equals(
+					"com/liferay/util/bean/PortletBeanLocatorUtil.class")) {
+
+				URL url = findResource("/");
+
+				String path = url.getPath();
+
+				path = path.substring(
+					0, path.length() - RESOURCE_PATH.length() - 1);
+
+				path = path.concat(name);
+
+				try {
+					return new URL("file", null, path);
+				}
+				catch (MalformedURLException murle) {
+				}
+			}
+
+			URL url = findResource(name);
+
+			if (url != null) {
+				return url;
+			}
+
+			return super.getResource(name);
+		}
+
+		@Override
+		public Class<?> loadClass(String name) throws ClassNotFoundException {
+			if (name.startsWith(_PACKAGE_PATH)) {
+				if (_classes.containsKey(name)) {
+					return _classes.get(name);
+				}
+
+				Class<?> clazz = super.findClass(name);
+
+				_classes.put(name, clazz);
+
+				return clazz;
+			}
+
+			return super.loadClass(name);
+		}
+
+		@Override
+		protected synchronized Class<?> loadClass(String name, boolean resolve)
+			throws ClassNotFoundException {
+
+			if (name.startsWith(_PACKAGE_PATH)) {
+				if (_classes.containsKey(name)) {
+					return _classes.get(name);
+				}
+
+				Class<?> clazz = super.findClass(name);
+
+				_classes.put(name, clazz);
+
+				return clazz;
+			}
+
+			return super.loadClass(name, resolve);
+		}
+
+		private final Map<String, Class<?>> _classes =
+			new ConcurrentHashMap<String, Class<?>>();
+		private final Map<String, URL> _urls =
+			new ConcurrentHashMap<String, URL>();
+
 	}
 
 	private static class PACLResourceLoader implements ResourceLoader {
@@ -171,9 +376,7 @@ public class PACLTestRule extends BaseTestRule<HotDeployEvent, Object> {
 		public Resource getResource(String location) {
 			ClassLoader classLoader = getClassLoader();
 
-			return new ClassPathResource(
-				PACLIntegrationJUnitTestRunner.RESOURCE_PATH + location,
-				classLoader);
+			return new ClassPathResource(RESOURCE_PATH + location, classLoader);
 		}
 
 		private final ClassLoader _classLoader;
