+++ portal-test-internal/src/com/liferay/portal/test/pacl-test.properties
